+++ PRTelegramBot/Models/InlineButtons/InlineSwitchInlineQueryCurrentChat.cs
@@ -1,0 +1,43 @@
+﻿using PRTelegramBot.Interfaces;
+
+namespace PRTelegramBot.Models.InlineButtons
+{
+    /// <summary>
+    /// Создает кнопку встроенной клавиатуры. При нажатии на кнопку в поле ввода текущего чата будет вставлено имя пользователя бота и указанный инлайн-запрос. Может быть пустым, в этом случае будет вставлено только имя пользователя бота.<br/><br/>Это предлагает быстрый способ для пользователя открыть вашего бота в инлайн-режиме в том же чате - подходит для выбора чего-либо из нескольких вариантов. Не поддерживается в каналах и для сообщений, отправленных от имени аккаунта Telegram Business.
+    /// </summary>
+    public class InlineSwitchInlineQueryCurrentChat : InlineBase, IInlineContent
+    {
+        #region Поля и свойства
+
+        /// <summary>
+        /// Если установлено, при нажатии на кнопку в поле ввода текущего чата будет вставлено имя пользователя бота и указанный инлайн-запрос. Может быть пустым, в этом случае будет вставлено только имя пользователя бота.<br/><br/>Это предлагает быстрый способ для пользователя открыть вашего бота в инлайн-режиме в том же чате - подходит для выбора чего-либо из нескольких вариантов. Не поддерживается в каналах и для сообщений, отправленных от имени аккаунта Telegram Business.
+        /// </summary>
+        public string SwitchInlineQueryCurrentChat { get; set; }
+
+        #endregion
+
+        #region IInlineContent
+
+        public object GetContent()
+        {
+            return SwitchInlineQueryCurrentChat;
+        }
+
+        #endregion
+
+        #region Конструкторы
+
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="buttonName">Название кнопки.</param>
+        /// <param name="switchInlineQueryCurrentChat">Если установлено, при нажатии на кнопку в поле ввода текущего чата будет вставлено имя пользователя бота и указанный инлайн-запрос. Может быть пустым, в этом случае будет вставлено только имя пользователя бота.<br/><br/>Это предлагает быстрый способ для пользователя открыть вашего бота в инлайн-режиме в том же чате - подходит для выбора чего-либо из нескольких вариантов. Не поддерживается в каналах и для сообщений, отправленных от имени аккаунта Telegram Business.</param>
+        public InlineSwitchInlineQueryCurrentChat(string buttonName, string switchInlineQueryCurrentChat)
+            : base(buttonName)
+        {
+            SwitchInlineQueryCurrentChat = switchInlineQueryCurrentChat;
+        }
+
+        #endregion
+    }
+}